IO2 impl. No StackOverflow example
