+++ src/main/scala/chapter_13/IO2.scala
@@ -1,6 +1,30 @@
 package chapter_13
 
-sealed trait IO2[A] {
+import chapter_11.Monad
+import chapter_11.MonadOps._
+import chapter_13.IO2._
+
+import scala.annotation.tailrec
+
+sealed trait IO2[A] { self =>
+
+  @tailrec
+  final def run: A = self match {
+    case Return(a) =>
+      a
+    case Suspend(resume) =>
+      resume()
+    case FlatMap(x, f) => x match {
+      case Return(a) =>
+        f(a).run
+      case Suspend(resume) =>
+        f(resume()).run
+      case FlatMap(x2, f2) =>
+        x2.flatMap(f2).flatMap(f).run
+    }
+
+  }
+
   def flatMap[B](f: A => IO2[B]): IO2[B] =
     FlatMap(this, f)
 
@@ -8,8 +32,29 @@ sealed trait IO2[A] {
     flatMap(f andThen (Return(_)))
 }
 
-case class Return[A](a: A) extends IO2[A]
+object IO2 {
 
-case class Suspend[A](resume: () => A) extends IO2[A]
+  implicit val IO2Monad = new Monad[IO2] {
+    def unit[A](a: => A): IO2[A] = Return(a)
 
-case class FlatMap[A, B](sub: IO2[A], k: A => IO2[B]) extends IO2[B]
+    def flatMap[A, B](fa: IO2[A])(f: A => IO2[B]) = fa flatMap f
+  }
+
+  case class Return[A](a: A) extends IO2[A]
+
+  case class Suspend[A](resume: () => A) extends IO2[A]
+
+  case class FlatMap[A, B](sub: IO2[A], k: A => IO2[B]) extends IO2[B]
+
+  def printLine(s: String): IO2[Unit] =
+    Suspend(() => Return(println(s)))
+
+}
+
+
+object IO2App extends App {
+
+  import IO2._
+
+  forever(printLine("Still going...")).run
+}
